requirements.txt
Django
celery
psycopg2
python-dateutil
Pillow
django-rest-framework
drf-nested-routers
google-auth
django-cockroachdb==3.1.*
gunicorn
git+https://github.com/AS207960/python-utils@85ce15b3e4b8a663efcf5b78d36cf44223dc2427
# git+https://github.com/TheEnbyperor/django-keycloak-auth@d0c11a4a5dec5dfc7730256b011bc86260dda5cf
ics
